feat: change sensor request and response entities
