--- internal/server/http/entities/sensor/request.go
@@ -1,11 +1,0 @@
-package sensor
-
-type SensorCreateRequest struct {
-	Status SensorStatus `json:"status"`
-	Type   string       `json:"type"`
-} // @Name SensorCreateRequest
-
-type SensorUpdateRequest struct {
-	Status SensorStatus `json:"status"`
-	Type   string       `json:"type"`
-} // @Name SensorUpdateRequest
